requirements.txt
flask==3.1.3
pillow==11.3.0
numpy==2.0.2
scipy==1.13.1
requests==2.32.3
gunicorn==23.0.0